Design revision 1 - Introduced GameStateChange
Original design depended on a mutable GameState and circular
references between GameState and Player, Square, etc. We feared the
mess this would cause when testing and debugging. This revision aims
for a better design by making the GameState object essentially
immutable, with players and squares submitting changes to it through
the GameStateChange object. This decouples the fixed state of the
game (GameState) from the changes made to it (GameStateChange). The
sequence and class diagrams have been revised to outline the new
design.

- Introduced GameStateChange
- Revised sequence diagrams
- Revised class diagrams