Implemented Card
In order to implement Card, we required the use of a deck data
structure, which is written up as a separate object. The Deck
represents the deck of chance and community chest cards by holding
a list of lambdas, which implement the changes that each card makes
to the GameState.

- Implemented Card according to class diagram
- Added the Deck data structure for keeping track of the chance
and community chest cards