Attempted to fix conservative_decision_maker... see below
It turns out that coding whether a player can build on a property
creates a lot of duplicate logic, because you have to basically
pre-apply changes to the state to see if you have enough money, etc.

I think it will be easier to just refactor the GameState to be
immutable, and every call to apply() will create a new GameState.
So you dont have to apply changes yourself in the DecisionMaker.

I'm going to give that a shot instead... wish me luck, hope I
dont break the whole game :P